// generated/src/app/api-models/letter-history.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class LetterHistory {
  seqLetterHist: number;
  letterId: string;
  letterIdSuffix: string;
  letterType: string;
  addresseeType: string;
  seqMembId: number;
  seqGroupId: number;
  seqProvId: number;
  seqVendId: number;
  printStatus: string;
  printDate: Date;
  printBatch: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  seqLtprtId: number;
  addresseeId: string;
  addresseeName: string;
  addresseeZipCode: string;
}